+++ src/commands/import-bookmarks.ts
@@ -1,13 +1,14 @@
 import * as vscode from 'vscode';
 import * as fs from 'fs';
 import {
-	ARE_YOU_SURE_IMPORT_MSG, 
-	ERROR_IMPORTING_MSG, 
-	NO_MSG, YES_MSG } from '../consts/messages';
+	BOOKMARK_IMPORT_QUESTION, 
+	BOOKMARK_ERR_IMPORTING, 
+	GENERIC_YES_ANSWER, GENERIC_NO_ANSWER 
+} from '../consts/messages';
 import { plainToClass } from 'class-transformer';
-import { GithubRepository } from '../models/github-repository';
+import { GithubRepositoryModel } from '../models/github-repository-model';
 import { openDialogOptions } from '../utils/dialog-options';
-import { CategoriesRepositories } from '../models/categories-repositories';
+import { CategoriesRepositoriesModel } from '../models/categories-repositories-model';
 import { IMPORT_BOOKMARKS } from '../consts/commands';
 import { inject, injectable } from 'inversify';
 import { Command } from './base/command';
@@ -29,7 +30,7 @@ export class ImportBookmarks implements Command {
 
 	async execute() {
 		if(!this.bookmarkManager.categoryRepositories) {
-			this.bookmarkManager.categoryRepositories = new CategoriesRepositories();
+			this.bookmarkManager.categoryRepositories = new CategoriesRepositoriesModel();
 		}
 
 		const hasRepositories = this.bookmarkManager
@@ -37,11 +38,11 @@ export class ImportBookmarks implements Command {
 
 		if(hasRepositories) {
 			vscode.window.showInformationMessage(
-				ARE_YOU_SURE_IMPORT_MSG,
-				...[YES_MSG, NO_MSG]
+				BOOKMARK_IMPORT_QUESTION,
+				...[GENERIC_YES_ANSWER, GENERIC_NO_ANSWER]
 			)
 			.then((answer) => {
-				if (answer === 'Yes') {
+				if (answer === GENERIC_YES_ANSWER) {
 					this.pickFolder();
 				}
 			});
@@ -51,38 +52,49 @@ export class ImportBookmarks implements Command {
 		}
 	}
 
-	async pickFolder() {
+	private async pickFolder() {
 		return await vscode.window
 		.showOpenDialog(openDialogOptions)
 			.then(fileUri => {
-				if (fileUri && fileUri[0]) {
-					fs.readFile(fileUri[0].fsPath, 'utf8' , (err, data) => {
-					if (err) {
-						console.error(err);
-						return;
-					}
+				try {
+					if (fileUri && fileUri[0]) {
+						fs.readFile(fileUri[0].fsPath, 'utf8' , (err, data) => {
+							if (err) {
+								console.error(err);
+								return;
+							}
 
-					this.validateImport(data);
-				});
-			}
+							this.validateAndImport(data);
+						});
+					}
+				}
+				catch (error) {
+					vscode.window.showErrorMessage(BOOKMARK_ERR_IMPORTING);
+				}
 		});
 	}
 	
-	async validateImport(data: string) {
-		try {
-			const parsedObject = JSON.parse(data);
-			const categoriesRepositories = plainToClass(CategoriesRepositories, parsedObject);
-			categoriesRepositories.repositories.forEach(parsedRepository => {
-				const repository  = plainToClass(GithubRepository, parsedRepository);
-				if(repository) {
-					repository.setCloneUrl();
-				}
-			});
+	private async validateAndImport(data: string) {
+		const parsedObject = JSON.parse(data);
+		const categoriesRepositories = plainToClass(CategoriesRepositoriesModel, parsedObject);
+
+		if(categoriesRepositories.categories.length === 0) {
+			vscode.window.showErrorMessage(BOOKMARK_ERR_IMPORTING);
+			return;
+		}
+
+		// Adding and updating objects in the data storage
+		categoriesRepositories.repositories.forEach(storedRepository => {
+			const repository  = plainToClass(GithubRepositoryModel, storedRepository);			
+			if(repository) {
+				// updating stored object
+				let index = categoriesRepositories.repositories
+				.indexOf(storedRepository);
+				categoriesRepositories!.repositories[index] = repository;
+			}
 
 			this.bookmarkManager.categoryRepositories = categoriesRepositories;
 			this.bookmarkManager.storeAndRefreshProvider();
-		} catch (error) {
-			vscode.window.showErrorMessage(ERROR_IMPORTING_MSG);
-		}
+		});		
 	}
 }