+++ src/models/categories-repositories.ts
@@ -1,8 +1,0 @@
-import { Category } from "./category";
-import { GithubRepository } from "./github-repository";
-
-// Storable object with Categories and Repositories
-export class CategoriesRepositories {
-	categories: Category[] = [];
-	repositories: GithubRepository[] = [];
-}